init project

+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -13,6 +13,10 @@ class CreateFailedJobsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('failed_jobs')) {
+            return;
+        }
+
         Schema::create('failed_jobs', function (Blueprint $table) {
             $table->id();
             $table->text('connection');
@@ -30,6 +34,10 @@ class CreateFailedJobsTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('failed_jobs')) {
+            return;
+        }
+
         Schema::dropIfExists('failed_jobs');
     }
 }
